app/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
influxdb-client
